+++ example/android/app/src/main/kotlin/ng/balanci/example/MainActivity.kt
@@ -1,0 +1,6 @@
+package ng.balanci.example
+
+import io.flutter.embedding.android.FlutterActivity
+
+class MainActivity: FlutterActivity() {
+}